requirements.txt
pandas>=2.0
numpy>=1.26
scikit-learn>=1.3
mlflow>=2.16
fastapi>=0.110
uvicorn[standard]>=0.23
click>=8.1
pydantic>=2.6
python-dotenv>=1.0
matplotlib>=3.8
requests>=2.31
azure-storage-blob>=12.19
